* Controllers split up * Logging/Configuration moved to Atom
