+++ Directory.Packages.props
@@ -12,11 +12,8 @@
     <PackageVersion Include="System.Reactive" Version="6.0.0" />
     <PackageVersion Include="Microsoft.Extensions.DependencyInjection" Version="$(FrameworkPackageVersion)" />
     <PackageVersion Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="$(FrameworkPackageVersion)"/>
-    <PackageVersion Include="Microsoft.Extensions.Configuration.Binder" Version="$(FrameworkPackageVersion)" />
-    <PackageVersion Include="Microsoft.Extensions.Configuration.Json" Version="$(FrameworkPackageVersion)" />
     <PackageVersion Include="Microsoft.Extensions.Logging" Version="$(FrameworkPackageVersion)" />
     <PackageVersion Include="Microsoft.Extensions.Logging.Abstractions" Version="$(FrameworkPackageVersion)" />
-    <PackageVersion Include="Microsoft.Extensions.Logging.Debug" Version="$(FrameworkPackageVersion)" />
     <PackageVersion Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="$(FrameworkPackageVersion)" />
     <PackageVersion Include="System.Threading.Tasks.Dataflow" Version="$(FrameworkPackageVersion)" />
 
@@ -37,13 +34,6 @@
 
     <!-- Misc dependencies -->
     <PackageVersion Include="CommandLineParser" Version="2.9.1" />
-    <PackageVersion Include="Serilog.Enrichers.Thread" Version="3.1.0" />
-    <PackageVersion Include="Serilog.Expressions" Version="4.0.0" />
-    <PackageVersion Include="Serilog.Extensions.Logging" Version="8.0.0" />
-    <PackageVersion Include="Serilog.Extensions.Logging.File" Version="3.0.0" />
-    <PackageVersion Include="Serilog.Settings.Configuration" Version="8.0.0" />
-    <PackageVersion Include="Serilog.Sinks.Console" Version="5.0.1" />
-    <PackageVersion Include="Serilog.Sinks.File" Version="5.0.0" />
 
     <!-- Automated Testing components -->
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
