+++ Starlog.UI/App.xaml.cs
@@ -4,6 +4,8 @@ global using System.Linq;
 global using System.Windows;
 global using Genius.Atom.Infrastructure;
 global using Genius.Atom.UI.Forms;
+using System.IO;
+using System.Reflection;
 using Genius.Starlog.Core;
 using Genius.Starlog.UI.AutoGridBuilders;
 using Genius.Starlog.UI.Console;
@@ -12,10 +14,8 @@ using Genius.Starlog.UI.Helpers;
 using Genius.Starlog.UI.Views;
 using Genius.Starlog.UI.Views.Comparison;
 using Genius.Starlog.UI.Views.LogSearchAndFiltering;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Serilog;
 
 namespace Genius.Starlog.UI;
 
@@ -38,11 +38,21 @@ public partial class App : Application
     {
         base.OnStartup(e);
 
+#if DEBUG
+        // This is necessary to manually resolve Atom.UI.Forms's dependencies,
+        // when referenced directly to dll, avoiding NuGet.
+        AppDomain.CurrentDomain.AssemblyResolve += (sender, args) => {
+            var proposedPath = Path.GetDirectoryName(args.RequestingAssembly!.Location) + "\\" + args.Name.Split(',')[0] + ".dll";
+            if (!File.Exists(proposedPath))
+                return null;
+            var assembly = Assembly.LoadFile(proposedPath);
+            return assembly;
+        };
+#endif
+
         var serviceCollection = new ServiceCollection();
 
-        var config = LoadConfiguration(serviceCollection);
-        ConfigureLogging(serviceCollection, config);
-        ConfigureServices(serviceCollection, config);
+        ConfigureServices(serviceCollection);
 
         ServiceProvider = serviceCollection.BuildServiceProvider();
         Atom.Data.Module.Initialize(ServiceProvider);
@@ -69,16 +79,20 @@ public partial class App : Application
         base.OnExit(e);
     }
 
-    private void ConfigureServices(IServiceCollection services, IConfiguration configuration)
+    private void ConfigureServices(IServiceCollection services)
     {
         Atom.Data.Module.Configure(services);
         Atom.Infrastructure.Module.Configure(services);
-        Atom.UI.Forms.Module.Configure(services, this);
+        var configuration = Atom.UI.Forms.Module.Configure(services, this);
         Starlog.Core.Module.Configure(services, configuration);
 
-        // Views, View models, View model factories, Controllers
-        services.AddSingleton<MainWindow>();
+        // Controllers
+        services.AddSingleton<IComparisonController, ComparisonController>();
+        services.AddSingleton<IConsoleController, ConsoleController>();
         services.AddSingleton<IMainController, MainController>();
+
+        // Views, View models, View model factories
+        services.AddSingleton<MainWindow>();
         services.AddSingleton<IViewModelFactory, ViewModelFactory>();
         services.AddSingleton<IMainViewModel, MainViewModel>();
         services.AddSingleton<IProfilesViewModel, ProfilesViewModel>();
@@ -100,31 +114,6 @@ public partial class App : Application
         services.AddTransient<IClipboardHelper, ClipboardHelper>();
         services.AddTransient<ILogArtifactsFormatter, LogArtifactsFormatter>();
         services.AddTransient<IConsoleParser, ConsoleParser>();
-    }
-
-    private static IConfiguration LoadConfiguration(ServiceCollection serviceCollection)
-    {
-        var builder = new ConfigurationBuilder()
-            .SetBasePath(AppDomain.CurrentDomain.BaseDirectory)
-            .AddJsonFile("appsettings.json", optional: true);
-        IConfiguration config = builder.Build();
-        serviceCollection.AddSingleton<IConfiguration>(config);
-        return config;
-    }
-
-    private static void ConfigureLogging(ServiceCollection services, IConfiguration configuration)
-    {
-        Log.Logger = new LoggerConfiguration()
-            .ReadFrom.Configuration(configuration)
-            .Enrich.WithThreadId()
-            .Enrich.WithComputed("SourceContextName", "Substring(SourceContext, LastIndexOf(SourceContext, '.') + 1)")
-            .CreateLogger();
-
-        // Framework:
-        services.AddLogging(x =>
-        {
-            x.AddSerilog();
-        });
     }
 
     private void Application_DispatcherUnhandledException(object sender, System.Windows.Threading.DispatcherUnhandledExceptionEventArgs e)
